+++ CHANGELOG.md
@@ -27,7 +27,10 @@
 - Implemented Summoning. (#583)
   - Added `canvas.tokens.performTokenPlacement` as a helper method to place an actor on the canvas.
 - Added a `[[lookup]]` enricher which can be used to dynamically display document data in text. (#759)
-- Added `potency` as a new ability modifier key to increase the potency of abilities. (#1350)
+- New ability modifier keys
+  - Added `forced.pull`, `forced.push`, and `forced.slide` to adjust the distance of forced movement abilities. (#1321)
+  - Added `keyword` to add keywords to abilities. (#1322)
+  - Added `potency` to adjust the potency of abilities. (#1350)
 
 ### Changed